consumer/requirements.txt
kafka-python==2.0.2
google-cloud-bigquery==3.11.4
google-auth==2.22.0